Turn croatian stemmer into a library, add unit tests
Croatian stemmer is represented as a class, StemmerHr, with two
public methods, `stem` and `stem_text`, and their croatian
equivalents `korjenuj` and `korjenuj_tekst`.

`stem_text` / `korjenuj_tekst` accepts text as a string and tokenizes
it with an alphanumeric tokenizer: all whitespaces, punctuation and
other characters are delimiters.

`stem` / `korjenuj` actually does what a stemmer should do: given a
token, it stems it and returns a stemmed term.